+++ src/utils/use-multi-file-auth-state-redis-db.ts
@@ -6,10 +6,13 @@ import {
   SignalDataTypeMap,
 } from '@whiskeysockets/baileys';
 
+import { CacheService } from '../api/services/cache.service';
 import { Logger } from '../config/logger.config';
-import { RedisCache } from '../libs/redis.client';
 
-export async function useMultiFileAuthStateRedisDb(cache: RedisCache): Promise<{
+export async function useMultiFileAuthStateRedisDb(
+  instanceName: string,
+  cache: CacheService,
+): Promise<{
   state: AuthenticationState;
   saveCreds: () => Promise<void>;
 }> {
@@ -17,7 +20,7 @@ export async function useMultiFileAuthStateRedisDb(cache: RedisCache): Promise<{
 
   const writeData = async (data: any, key: string): Promise<any> => {
     try {
-      return await cache.writeData(key, data);
+      return await cache.hSet(instanceName, key, data);
     } catch (error) {
       return logger.error({ localError: 'writeData', error });
     }
@@ -25,16 +28,16 @@ export async function useMultiFileAuthStateRedisDb(cache: RedisCache): Promise<{
 
   const readData = async (key: string): Promise<any> => {
     try {
-      return await cache.readData(key);
+      return await cache.hGet(instanceName, key);
     } catch (error) {
-      logger.error({ readData: 'writeData', error });
+      logger.error({ localError: 'readData', error });
       return;
     }
   };
 
   const removeData = async (key: string) => {
     try {
-      return await cache.removeData(key);
+      return await cache.hDelete(instanceName, key);
     } catch (error) {
       logger.error({ readData: 'removeData', error });
     }